backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.3
scikit-learn==1.3.2
